Refactor <SidePanel /> as functional component
