+++ resources/assets/js/components/dashboard/SidePanel.js
@@ -1,16 +1,13 @@
 import React, { Component } from "react";
-import ReactDOM from "react-dom";
-import { Col, Row } from "react-bootstrap";
-import NavItem from "./NavItem";
+import NavItem from "../NavItem";
 
-export default class SidePanel extends Component {
-	render() {
-		return (
-			<div className="dashboard-side-panel">
-				<NavItem name="Dashboard" link="/" />
-				<NavItem name="Exercises" link="/exercise" />
-				{/* <NavItem name="Muscle Map" link="/musclemap" /> */}
-			</div>
-		);
-	}
+export function SidePanel()
+{
+	return (
+		<div className="dashboard-side-panel">
+			<NavItem name="Dashboard" link="/" />
+			<NavItem name="Exercises" link="/exercise" />
+			{/* <NavItem name="Muscle Map" link="/musclemap" /> */}
+		</div>
+	);
 }
